# Prioritise and reduce the amount of events decrypted on application startup