--- a/src/models/room.js
+++ b/src/models/room.js
@@ -228,12 +228,6 @@ function pendingEventsKey(roomId) {
 
 utils.inherits(Room, EventEmitter);
 
-function shouldAttemptDecryption(event) {
-    return event.isEncrypted()
-        && !event.isBeingDecrypted()
-        && event.getClearContent() === null
-}
-
 Room.prototype.decryptCriticalEvents = function() {
     const readReceiptEventId = this.getEventReadUpTo(this._client.getUserId(), true);
     const events = this.getLiveTimeline().getEvents();
@@ -243,7 +237,7 @@ Room.prototype.decryptCriticalEvents = function() {
 
     const decryptionPromises = events
         .slice(readReceiptTimelineIndex)
-        .filter(event => shouldAttemptDecryption(event))
+        .filter(event => event.shouldAttemptDecryption())
         .reverse()
         .map(event => event.attemptDecryption(this._client._crypto, true));
 
@@ -255,7 +249,7 @@ Room.prototype.decryptAllEvents = function() {
         .getUnfilteredTimelineSet()
         .getLiveTimeline()
         .getEvents()
-        .filter(event => shouldAttemptDecryption(event))
+        .filter(event => event.shouldAttemptDecryption())
         .reverse()
         .map(event => event.attemptDecryption(this._client._crypto, true));