--- a/src/sync.js
+++ b/src/sync.js
@@ -701,31 +701,7 @@ SyncApi.prototype._syncFromCache = async function(savedSync) {
     };
 
     try {
-        await this._processSyncResponse(syncEventData, data, false);
-
-        const breadcrumbs = this.client.store.getAccountData("im.vector.setting.breadcrumbs");
-        const breadcrumbsRooms = breadcrumbs?.getContent().recent_rooms || [];
-
-        this.client.getRooms().forEach(room => {
-            const readReceiptEventId = room.getEventReadUpTo(this.client.getUserId(), true);
-
-            const events = room.getLiveTimeline().getEvents();
-            const isInBreadcrumb = breadcrumbsRooms.includes(room.roomId);
-            const readReceiptTimelineIndex = events.findIndex(matrixEvent => {
-                return matrixEvent.event.event_id === readReceiptEventId
-            });
-            const decryptFromIndex = isInBreadcrumb
-                ? 0
-                : readReceiptTimelineIndex
-
-            events
-                .slice(decryptFromIndex)
-                .forEach(event => {
-                    if (event.isEncrypted() && !event.isBeingDecrypted()) {
-                        event.attemptDecryption(this.client._crypto, true);
-                    }
-                });
-        });
+        await this._processSyncResponse(syncEventData, data);
     } catch (e) {
         logger.error("Error processing cached sync", e.stack || e);
     }
@@ -966,7 +942,7 @@ SyncApi.prototype._onSyncError = function(err, syncOptions) {
  * @param {Object} data The response from /sync
  */
 SyncApi.prototype._processSyncResponse = async function(
-    syncEventData, data, decrypt = true
+    syncEventData, data
 ) {
     const client = this.client;
     const self = this;
@@ -1182,10 +1158,11 @@ SyncApi.prototype._processSyncResponse = async function(
     await utils.promiseMapSeries(joinRooms, async function(joinObj) {
         const room = joinObj.room;
         const stateEvents = self._mapSyncEventsFormat(joinObj.state, room);
-        const timelineEvents = self._mapSyncEventsFormat(joinObj.timeline, room, decrypt);
+        const timelineEvents = self._mapSyncEventsFormat(joinObj.timeline, room, false);
         const ephemeralEvents = self._mapSyncEventsFormat(joinObj.ephemeral);
         const accountDataEvents = self._mapSyncEventsFormat(joinObj.account_data);
 
+        const encrypted = client.isRoomEncrypted(room.roomId);
         // we do this first so it's correct when any of the events fire
         if (joinObj.unread_notifications) {
             room.setUnreadNotificationCount(
@@ -1196,7 +1173,6 @@ SyncApi.prototype._processSyncResponse = async function(
             // bother setting it here. We trust our calculations better than the
             // server's for this case, and therefore will assume that our non-zero
             // count is accurate.
-            const encrypted = client.isRoomEncrypted(room.roomId);
             if (!encrypted
                 || (encrypted && room.getUnreadNotificationCount('highlight') <= 0)) {
                 room.setUnreadNotificationCount(
@@ -1318,6 +1294,11 @@ SyncApi.prototype._processSyncResponse = async function(
         });
 
         room.updateMyMembership("join");
+
+        // Decrypt only the last message in all rooms to make sure we can generate a preview
+        // And decrypt all events after the recorded read receipt to ensure an accurate
+        // notification count
+        room.decryptCriticalEvents();
     });
 
     // Handle leaves (e.g. kicked rooms)
@@ -1546,7 +1527,6 @@ SyncApi.prototype._mapSyncEventsFormat = function(obj, room, decrypt = true) {
     if (!obj || !utils.isArray(obj.events)) {
         return [];
     }
-    const mapper = this.client.getEventMapper();
     const mapper = this.client.getEventMapper({ decrypt });
     return obj.events.map(function(e) {
         if (room) {