Related to vector-im/element-web#13266
To review alongside matrix-org/matrix-react-sdk#5980

When loading element, it will either read from cache the entire `/sync` blob that has been persisted or will query that information from Synapse.
The idea is to implement a priority queue that will cap the maximum of events that will be decrypted on application startup. Everything else will be lazily decrypted when a user views a room

At the moment, from clicking on a tile to having all messages displayed on screen it takes roughly a second. With lazy decryption it will approximately take 2 seconds but the most recent messages will start to appear on screen from the first second.

# Critical events

This pull request brings the concept of critical events. These are the events that a client requires to have decrypted to ensure it can run and provide all the required features to the user

At the moment a critical event is an event that falls under the following category
* Last event of a room (to ensure previews are working)
* All events that occured after the read marker (to ensure correct-ness in notification count)

Let me know if you can think about anything else that would benefit from being decrypted ahead of time

In addition to that, I am using breadcrumb rooms to detect which rooms a user might be likely to interact with again in the future and proactively decrypt the live timeline. 
This could be scrapped if we wanted to further improve app start performance at the cost of a few layout shift as highlighted in [this video](https://github.com/matrix-org/matrix-js-sdk/pull/1684#issuecomment-833583200)

# Benchmark

| # of e2ee rooms  | load time (before) | load time (after) | % difference | 
| ---------------- | ------------------ | ----------------- | ------------ |
| 250 | 4mn50 | 31s | **89%** |
| 43 | 37s | 12s | **67%** |

These recordings are taken on an existing session by looking at the time between a hard refresh until the console stops spamming "decrypted events" messages by the thousands.

Please share with me your recording and I will add them to the table above, the code is [up and running on a netlify preview](https://deploy-preview-5980--matrix-react-sdk.netlify.app/)
If you want to know the number of encrypted rooms in your account you can run the following snippet in your terminal

```javascript
Object.values(mxMatrixClientPeg.get().store.rooms).filter(room => {
  return room._client.isRoomEncrypted(room.roomId)
}).length;
```