@dbkr 2021-05-06
in general this looks like a good approach I think