@gsouquet 2021-05-06
> Do you have any data on how much performance improves with this change in isolation?

I have just done a benchmark using the same account. It takes down the load time from **4 minutes 50 seconds** down to **30 seconds**. During which the application is still usable, but with a degraded experience (sometimes the room spinner is missing, ...)

And below is a screen recording of the experience when we load a room where we need to lazily decrypt events


https://user-images.githubusercontent.com/769871/117318458-56ca1d80-ae82-11eb-9141-4c5eb226832e.mov