@jryans 2021-05-07
Ah yeah, as mentioned before, it would be good to have some kind of automated to verify rooms still decrypt as expected, either here or in the React layer (whichever makes more sense).